refactor(account): Create separated files for each endpoint
